Refactor tvl on backend (#2068)
