+++ packages/backend/src/api/controllers/tvl/charts.ts
@@ -1,5 +1,5 @@
 import {
-  DetailedTvlApiChartPoint,
+  TokenTvlApiChartPoint,
   TvlApiChartPoint,
   UnixTime,
 } from '@l2beat/shared-pure'
@@ -7,7 +7,7 @@ import {
 import { ReportRecord } from '../../../peripherals/database/ReportRepository'
 import { asNumber } from './asNumber'
 
-interface DetailedBalanceInTime {
+interface BalanceInTime {
   timestamp: UnixTime
   usdTvl: bigint
   ethTvl: bigint
@@ -19,17 +19,17 @@ interface DetailedBalanceInTime {
   ethNmv: bigint
 }
 
-export function addMissingTimestamps(
-  points: TvlApiChartPoint[],
+export function addTokenMissingTimestamps(
+  points: TokenTvlApiChartPoint[],
   hours: number,
-): TvlApiChartPoint[] {
+): TokenTvlApiChartPoint[] {
   if (points.length === 0) return []
   const [min] = points[0]
   const [max] = points[points.length - 1]
   const timestampValues = new Map(
     points.map(([t, v1, v2]) => [t.toString(), [v1, v2]]),
   )
-  const allPoints: TvlApiChartPoint[] = []
+  const allPoints: TokenTvlApiChartPoint[] = []
   for (
     let timestamp = min;
     timestamp.lte(max);
@@ -46,17 +46,17 @@ export function addMissingTimestamps(
   return allPoints
 }
 
-export function addDetailedMissingTimestamps(
-  points: DetailedTvlApiChartPoint[],
+export function addMissingTimestamps(
+  points: TvlApiChartPoint[],
   hours: number,
-): DetailedTvlApiChartPoint[] {
+): TvlApiChartPoint[] {
   if (points.length === 0) return []
   const [min] = points[0]
   const [max] = points[points.length - 1]
   const timestampValues = new Map(
     points.map(([t, ...values]) => [t.toString(), values]),
   )
-  const allPoints: DetailedTvlApiChartPoint[] = []
+  const allPoints: TvlApiChartPoint[] = []
   for (
     let timestamp = min;
     timestamp.lte(max);
@@ -99,25 +99,25 @@ export function getChartPoints(
   hours: number,
   decimals: number,
   usdFirst = false,
-): TvlApiChartPoint[] {
-  const existing: TvlApiChartPoint[] = balances.map((b) => {
+): TokenTvlApiChartPoint[] {
+  const existing: TokenTvlApiChartPoint[] = balances.map((b) => {
     const usd = asNumber(b.usd, 2)
     const asset = asNumber(b.asset, decimals)
     return usdFirst ? [b.timestamp, usd, asset] : [b.timestamp, asset, usd]
   })
-  return addMissingTimestamps(existing, hours)
+  return addTokenMissingTimestamps(existing, hours)
 }
 
 export function covertBalancesToChartPoints(
-  balancesInTime: DetailedBalanceInTime[],
+  balancesInTime: BalanceInTime[],
   resolutionInHours: number,
   decimals: number,
   usdFirst = false,
-): DetailedTvlApiChartPoint[] {
+): TvlApiChartPoint[] {
   const USD_DECIMALS = 2
 
-  const existingBalanceChartPoints: DetailedTvlApiChartPoint[] =
-    balancesInTime.map((bit) => {
+  const existingBalanceChartPoints: TvlApiChartPoint[] = balancesInTime.map(
+    (bit) => {
       const [usdTvl, usdCbv, usdEbv, usdNmv] = [
         bit.usdTvl,
         bit.usdCbv,
@@ -155,10 +155,8 @@ export function covertBalancesToChartPoints(
             usdEbv,
             usdNmv,
           ]
-    })
-
-  return addDetailedMissingTimestamps(
-    existingBalanceChartPoints,
-    resolutionInHours,
+    },
   )
+
+  return addMissingTimestamps(existingBalanceChartPoints, resolutionInHours)
 }
