+++ packages/backend/src/api/routers/TvlRouter.ts
@@ -8,15 +8,29 @@ import {
 } from '@l2beat/shared-pure'
 import { z } from 'zod'
 
-import { DetailedTvlController } from '../controllers/tvl/DetailedTvlController'
 import { TvlController } from '../controllers/tvl/TvlController'
 import { withTypedContext } from './types'
 
-export function createTvlRouter(
-  tvlController: TvlController,
-  detailedTvlController: DetailedTvlController,
-) {
+export function createTvlRouter(tvlController: TvlController) {
   const router = new Router()
+
+  router.get('/api/tvl', async (ctx) => {
+    const tvlResult = await tvlController.getTvlApiResponse()
+
+    if (tvlResult.result === 'error') {
+      if (tvlResult.error === 'DATA_NOT_FULLY_SYNCED') {
+        ctx.status = 422
+      }
+
+      if (tvlResult.error === 'NO_DATA') {
+        ctx.status = 404
+      }
+
+      return
+    }
+
+    ctx.body = tvlResult.data
+  })
 
   router.get(
     '/api/tvl/aggregate',
@@ -31,7 +45,7 @@ export function createTvlRouter(
         const projectSlugs = ctx.query.projectSlugs
 
         const tvlProjectsResponse =
-          await detailedTvlController.getDetailedAggregatedApiResponse(
+          await tvlController.getAggregatedTvlApiResponse(
             projectSlugs.split(',').map((slug) => slug.trim()),
           )
 
@@ -54,73 +68,6 @@ export function createTvlRouter(
     ),
   )
 
-  router.get('/api/tvl', async (ctx) => {
-    const tvlResponse = await tvlController.getTvlApiResponse()
-    if (tvlResponse.result === 'error') {
-      if (tvlResponse.error === 'DATA_NOT_FULLY_SYNCED') {
-        ctx.status = 422
-      }
-
-      if (tvlResponse.error === 'NO_DATA') {
-        ctx.status = 404
-      }
-
-      return
-    }
-
-    ctx.body = tvlResponse.data
-  })
-
-  router.get(
-    '/api/projects/:projectId/tvl/assets/:assetId',
-    withTypedContext(
-      z.object({
-        params: z.object({
-          projectId: branded(z.string(), ProjectId),
-          assetId: branded(z.string(), AssetId),
-        }),
-      }),
-      async (ctx) => {
-        const { projectId, assetId } = ctx.params
-        const chartResponse = await tvlController.getProjectAssetChart(
-          projectId,
-          assetId,
-        )
-        if (chartResponse.result === 'error') {
-          if (chartResponse.error === 'NO_DATA') {
-            ctx.status = 404
-          }
-
-          if (chartResponse.error === 'INVALID_PROJECT_OR_ASSET') {
-            ctx.status = 400
-          }
-
-          return
-        }
-        ctx.body = chartResponse.data
-      },
-    ),
-  )
-
-  router.get('/api/detailed-tvl', async (ctx) => {
-    const detailedTvlResult =
-      await detailedTvlController.getDetailedTvlApiResponse()
-
-    if (detailedTvlResult.result === 'error') {
-      if (detailedTvlResult.error === 'DATA_NOT_FULLY_SYNCED') {
-        ctx.status = 422
-      }
-
-      if (detailedTvlResult.error === 'NO_DATA') {
-        ctx.status = 404
-      }
-
-      return
-    }
-
-    ctx.body = detailedTvlResult.data
-  })
-
   router.get(
     '/api/projects/:projectId/tvl/chains/:chainId/assets/:assetId/types/:assetType',
 
@@ -136,34 +83,33 @@ export function createTvlRouter(
       async (ctx) => {
         const { assetId, chainId, assetType, projectId } = ctx.params
 
-        const detailedAssetData =
-          await detailedTvlController.getDetailedAssetTvlApiResponse(
-            projectId,
-            ChainId(+chainId),
-            assetId,
-            assetType,
-          )
+        const assetData = await tvlController.getAssetTvlApiResponse(
+          projectId,
+          ChainId(+chainId),
+          assetId,
+          assetType,
+        )
 
-        if (detailedAssetData.result === 'error') {
-          if (detailedAssetData.error === 'NO_DATA') {
+        if (assetData.result === 'error') {
+          if (assetData.error === 'NO_DATA') {
             ctx.status = 404
           }
 
-          if (detailedAssetData.error === 'INVALID_PROJECT_OR_ASSET') {
+          if (assetData.error === 'INVALID_PROJECT_OR_ASSET') {
             ctx.status = 400
           }
 
           return
         }
 
-        ctx.body = detailedAssetData.data
+        ctx.body = assetData.data
       },
     ),
   )
 
   router.get('/api/project-assets-breakdown', async (ctx) => {
     const projectAssetsBreakdown =
-      await detailedTvlController.getProjectTokenBreakdownApiResponse()
+      await tvlController.getProjectTokenBreakdownApiResponse()
 
     if (projectAssetsBreakdown.result === 'error') {
       if (projectAssetsBreakdown.error === 'NO_DATA') {
