+++ packages/backend/src/peripherals/database/ReportRepository.ts
@@ -93,62 +93,7 @@ export class ReportRepository extends BaseRepository {
     return knex('reports').delete()
   }
 
-  /**
-   * To be removed along old TVL api
-   * @deprecated
-   */
-  async getDailyByProjectAndAsset(
-    projectId: ProjectId,
-    assetId: AssetId,
-  ): Promise<ReportRecord[]> {
-    const knex = await this.knex()
-    const rows = await this._getByProjectAndAssetQuery(knex, projectId, assetId)
-      .andWhereRaw(`extract(hour from unix_timestamp) = 0`)
-      .whereIn('chain_id', [...ChainId.getAll()])
-      .whereIn('report_type', ['EBV', 'CBV', 'NMV'])
-
-    return rows.map(toRecord)
-  }
-
-  /**
-   * To be removed along old TVL api
-   * @deprecated
-   */
-  async getHourlyByProjectAndAsset(
-    projectId: ProjectId,
-    assetId: AssetId,
-    from: UnixTime,
-  ): Promise<ReportRecord[]> {
-    const knex = await this.knex()
-    const rows = await this._getByProjectAndAssetQuery(knex, projectId, assetId)
-      .andWhere('unix_timestamp', '>=', from.toDate())
-      .whereIn('chain_id', [...ChainId.getAll()])
-      .whereIn('report_type', ['EBV', 'CBV', 'NMV'])
-
-    return rows.map(toRecord)
-  }
-
-  /**
-   * To be removed along old TVL api
-   * @deprecated
-   */
-  async getSixHourlyByProjectAndAsset(
-    projectId: ProjectId,
-    assetId: AssetId,
-    from: UnixTime,
-  ): Promise<ReportRecord[]> {
-    const knex = await this.knex()
-    const rows = await this._getByProjectAndAssetQuery(knex, projectId, assetId)
-      .andWhereRaw(`extract(hour from "unix_timestamp") % 6 = 0`)
-      .andWhere('unix_timestamp', '>=', from.toDate())
-      .whereIn('chain_id', [...ChainId.getAll()])
-      .whereIn('report_type', ['EBV', 'CBV', 'NMV'])
-
-    return rows.map(toRecord)
-  }
-
-  // Detailed asset TVL
-  async getHourlyForDetailed(
+  async getHourly(
     projectId: ProjectId,
     chainId: ChainId,
     assetId: AssetId,
@@ -168,7 +113,7 @@ export class ReportRepository extends BaseRepository {
     return rows.map(toRecord)
   }
 
-  async getSixHourlyForDetailed(
+  async getSixHourly(
     projectId: ProjectId,
     chainId: ChainId,
     assetId: AssetId,
@@ -189,7 +134,7 @@ export class ReportRepository extends BaseRepository {
     return rows.map(toRecord)
   }
 
-  async getDailyForDetailed(
+  async getDaily(
     projectId: ProjectId,
     chainId: ChainId,
     assetId: AssetId,
