+++ packages/config/src/layer2s/types/Layer2.ts
@@ -103,8 +103,6 @@ export interface Layer2Config {
   escrows: ProjectEscrow[]
   /** API parameters used to get transaction count */
   transactionApi?: Layer2TransactionApi
-  /** Text displayed on the TVL number hover */
-  tvlTooltip?: string
   /** Configuration for getting state updates and batch submission */
   liveness?: Layer2LivenessConfig
 }
