+++ packages/frontend/src/pages/Page.ts
@@ -1,7 +1,7 @@
 import {
   ActivityApiResponse,
-  DetailedTvlApiResponse,
   ProjectAssetsBreakdownApiResponse,
+  TvlApiResponse,
   VerificationStatus,
 } from '@l2beat/shared-pure'
 import { ReactElement } from 'react'
@@ -30,7 +30,7 @@ export interface PageMetadata {
 }
 
 export interface PagesData {
-  tvlApiResponse: DetailedTvlApiResponse
+  tvlApiResponse: TvlApiResponse
   activityApiResponse: ActivityApiResponse | undefined
   verificationStatus: VerificationStatus
   tvlBreakdownApiResponse: ProjectAssetsBreakdownApiResponse | undefined
