+++ packages/frontend/src/pages/scaling/projects/props/getProjectHeader.ts
@@ -1,12 +1,11 @@
 import { Layer2, ProjectLinks } from '@l2beat/config'
 import {
   ActivityApiResponse,
-  DetailedTvlApiCharts,
-  DetailedTvlApiResponse,
+  TvlApiCharts,
+  TvlApiResponse,
 } from '@l2beat/shared-pure'
 
 import { Config } from '../../../../build/config'
-import { TvlStats } from '../../../../components/header/TvlSummary'
 import { formatLargeNumber } from '../../../../utils'
 import { getTpsDaily } from '../../../../utils/activity/getTpsDaily'
 import { getTpsWeeklyChange } from '../../../../utils/activity/getTpsWeeklyChange'
@@ -21,12 +20,12 @@ import { ProjectHeaderProps } from '../view/ProjectHeader'
 export function getProjectHeader(
   project: Layer2,
   config: Config,
-  tvlApiResponse: DetailedTvlApiResponse,
+  tvlApiResponse: TvlApiResponse,
   activityApiResponse?: ActivityApiResponse,
 ): ProjectHeaderProps {
   const apiProject = tvlApiResponse.projects[project.id.toString()]
 
-  const getDetailed = (chart: DetailedTvlApiCharts | undefined) => {
+  const getDetailed = (chart: TvlApiCharts | undefined) => {
     const { parts, partsWeeklyChange } = getDetailedTvlWithChange(chart)
     return {
       tvl: parts.tvl,
@@ -62,14 +61,13 @@ export function getProjectHeader(
     icon: `/icons/${project.display.slug}.png`,
     title: project.display.name,
     titleLength: getTitleLength(project.display.name),
-    tvlStats: getTvlStats(
-      project,
-      tvl,
-      tvlWeeklyChange,
+    tvlStats: {
+      tvlChange: tvlWeeklyChange,
+      tvl: project.config.escrows.length > 0 ? tvl : 0,
       canonical,
       external,
       native,
-    ),
+    },
     tpsDaily: tpsDaily?.toFixed(2) ?? '',
     tpsWeeklyChange,
     transactionMonthlyCount:
@@ -137,39 +135,4 @@ function getLinks(links: ProjectLinks) {
   ] as const
 
   return items.filter((link) => link.links.length > 0)
-}
-
-function getTvlStats(
-  project: Layer2,
-  tvl: number,
-  tvlWeeklyChange: string,
-  canonical: number | undefined,
-  external: number | undefined,
-  native: number | undefined,
-): TvlStats | undefined {
-  // TODO(radomski): This is solution is really janky and can be improved once
-  // we deprecate the usage of the original endpoint to fetch data for the frontend.
-  const parts = [
-    project.config.escrows.length > 0 ? tvl : 0,
-    canonical ?? 0,
-    external ?? 0,
-    native ?? 0,
-  ]
-
-  if (parts.every((x) => notUndefined(x))) {
-    const ps = parts.filter(notUndefined)
-    return {
-      tvlChange: tvlWeeklyChange,
-      tvl: ps[0],
-      canonical: ps[1],
-      external: ps[2],
-      native: ps[3],
-    }
-  }
-
-  return undefined
-}
-
-function notUndefined<T>(x: T | undefined): x is T {
-  return x !== undefined
 }
