+++ packages/frontend/src/utils/getIncludedProjects.ts
@@ -1,12 +1,8 @@
-import { DetailedTvlApiResponse, ProjectId } from '@l2beat/shared-pure'
+import { ProjectId, TvlApiResponse } from '@l2beat/shared-pure'
 
 export function getIncludedProjects<
   T extends { id: ProjectId; isUpcoming?: boolean; type: 'bridge' | 'layer2' },
->(
-  projects: T[],
-  tvlApiResponse: DetailedTvlApiResponse,
-  buildAllProjectPages = false,
-) {
+>(projects: T[], tvlApiResponse: TvlApiResponse, buildAllProjectPages = false) {
   if (buildAllProjectPages) {
     return projects
   }
