+++ packages/shared-pure/src/types/api/index.ts
@@ -1,4 +1,4 @@
 export * from './ActivityApi'
-export * from './DetailedTvlApi'
+export * from './TokenTvlApi'
 export * from './TvlApi'
 export * from './VerificationStatus'
